Dedupe ST CollectHeart
