+++ config/symbols.us.stcen.txt
@@ -4,6 +4,7 @@ g_InitializeData0 = 0x80180410;
 g_InitializeEntityData0 = 0x8018041C;
 UNK_Invincibility0 = 0x801807B8;
 c_GoldPrizes = 0x80180F58;
+c_HeartPrizes = 0x80180FE8;
 g_Rooms = 0x801812D4;
 g_UnkPrimHelperRot = 0x8018D5F8;
 EntityBackgroundBlock = 0x8018D600;
