+++ config/symbols.us.stno3.txt
@@ -12,6 +12,7 @@ g_eBreakableBlendModes = 0x80180E90;
 g_TrapDoorFlag = 0x8018123C;
 UNK_Invincibility0 = 0x80181CC0;
 c_GoldPrizes = 0x80182460;
+c_HeartPrizes = 0x801824F0;
 g_Rooms = 0x80183CC4;
 g_UnkPrimHelperRot = 0x801B77CC;
 EntityCavernDoorVase = 0x801B77D4;
